pyproject.toml
[project]
name = "tumor-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=1.3.0",
    "black>=23.3.0",
    "captum>=0.6.0",
    "flwr>=1.7.0",
    "grad-cam>=1.4.0",
    "ipykernel>=6.23.0",
    "kagglehub>=0.4.2",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.14.0",
    "pytest>=7.3.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.13.0",
    "timm>=0.9.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
]
